+++ attributes/default.rb
@@ -33,7 +33,6 @@ default['packer']['checksum'] = node['packer']['checksums'][filename]
 
 default['packer']['install_method'] = 'binary'
 default['packer']['install_dir'] = '/usr/local/bin'
-default['packer']['source_revision'] = 'master'
 
 # Windows attribute
 if platform_family?('windows')