+++ attributes/source.rb
@@ -1,0 +1,2 @@
+default['packer']['source_revision'] = 'master'
+default['packer']['source_repo_url'] = 'https://github.com/mitchellh/packer.git'