+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This is a changelog for the sbp_packer cookbook.
   * Starting the change log here, since none existed before
   * Updated for packer 0.9.0
   * added minitest-handler dependency for run-time assertions
+  * Added node['packer']['source_repo_url'] to contain the URL for the packer repo during a source install
 
 
 ------------------------------------